Renamed files with components with PaskalCase
