+++ src/componentsBasic/home/homeSections/homeInfo/HomeInfo.scss
