+++ src/componentsUniversal/button/button.js
@@ -1,4 +1,4 @@
-import './button.scss'
+import './Button.scss'
 
 const Button = ({text, className=''}) => {
     return (
