+++ src/componentsUniversal/footer/footer.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import "./footer.scss";
+import "./Footer.scss";
 
 const Footer = () => {
   return (
